abstract Parser class
