+++ src/main/java/Duke.java
@@ -4,10 +4,17 @@ import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.*;
 public class Duke {
+
+    private Ui newUser;
+    private Storage data;
+
+    public Duke(String filePath) throws IOException {
+        newUser = new Ui();
+        data = new Storage(newUser.getTaskArr(), filePath);
+    }
+
     public static void main(String[] args) throws IOException {
-        Ui newUser = new Ui();
-        newUser.userInput();
-        Storage.saveFile(newUser.getTaskArr(), "duke.txt");
+        new Duke("duke.txt");
     }
 
 }
